+++ src/layer/x86/unaryop_x86.h
@@ -14,6 +14,11 @@ public:
     UnaryOp_x86();
 
     virtual int forward_inplace(Mat& bottom_top_blob, const Option& opt) const;
+
+protected:
+#if NCNN_BF16
+    int forward_inplace_bf16s(Mat& bottom_top_blob, const Option& opt) const;
+#endif
 };
 
 } // namespace ncnn
